events: Clean up unused functions onChangeValue: Separate logic, write test
