--- src/events/createOnDragStart.js
@@ -1,8 +1,0 @@
-export const dataKey = 'text'
-const createOnDragStart =
-  (name, value) =>
-    event => {
-      event.dataTransfer.setData(dataKey, value)
-    }
-
-export default createOnDragStart
