--- src/events/createOnDrop.js
@@ -1,8 +1,0 @@
-import { dataKey } from './createOnDragStart'
-const createOnDrop =
-  (name, change) =>
-    event => {
-      change(event.dataTransfer.getData(dataKey))
-      event.preventDefault()
-    }
-export default createOnDrop
