--- src/events/createOnFocus.js
@@ -1,4 +1,0 @@
-const createOnFocus =
-  (name, focus) =>
-    () => focus(name)
-export default createOnFocus
